CMS-4638 Media Types - Make FileUploader in NewContentDialog work
- created abstract class Uploader
- created inheritors FileUploader and ContentUploader that upload to blob/upload and content/createMedia
- created ImageUploader extending ContentUploader that renders the response
- also updated usages in Image.ts, ImageSelector.ts, ImageUploadDialog.ts
- additionally used ImageUploadDialog which is the same as ImageSelectorUploadDialog.ts in ImageSelector.ts
