+++ modules/wem-admin-ui/src/main/resources/web/admin/common/js/content/form/inputtype/image/ImageUploadDialog.ts
@@ -1,8 +1,14 @@
 module api.content.form.inputtype.image {
 
+    import Content = api.content.Content;
+    import FileUploadedEvent = api.ui.uploader.FileUploadedEvent;
+    import FileUploadStartedEvent = api.ui.uploader.FileUploadStartedEvent;
+    import FileUploadProgressEvent = api.ui.uploader.FileUploadProgressEvent;
+    import FileUploadCompleteEvent = api.ui.uploader.FileUploadCompleteEvent;
+
     export class ImageUploadDialog extends api.ui.dialog.ModalDialog {
 
-        private imageUploader: api.ui.uploader.ImageUploader;
+        private imageUploader: api.content.ImageUploader;
 
         constructor() {
             super({
@@ -15,9 +21,8 @@ module api.content.form.inputtype.image {
             description.getEl().setInnerHtml("Images uploaded will be added as children to this content, you may move them later if desired");
             this.appendChildToContentPanel(description);
 
-            this.imageUploader = new api.ui.uploader.ImageUploader({
+            this.imageUploader = new api.content.ImageUploader({
                 name: 'image-selector-upload-dialog',
-                url: api.util.UriHelper.getRestUri("blob/upload"),
                 showButtons: false,
                 showResult: false
             });
@@ -47,12 +52,44 @@ module api.content.form.inputtype.image {
             super.open();
         }
 
-        onImageUploaded(listener: (event: api.ui.uploader.FileUploadedEvent) => void) {
+        onUploadStarted(listener: (event: FileUploadStartedEvent<Content>) => void) {
+            this.imageUploader.onUploadStarted(listener);
+        }
+
+        unUploadStarted(listener: (event: FileUploadStartedEvent<Content>) => void) {
+            this.imageUploader.unUploadStarted(listener);
+        }
+
+        onUploadProgress(listener: (event: FileUploadProgressEvent<Content>) => void) {
+            this.imageUploader.onUploadProgress(listener);
+        }
+
+        unUploadProgress(listener: (event: FileUploadProgressEvent<Content>) => void) {
+            this.imageUploader.unUploadProgress(listener);
+        }
+
+        onImageUploaded(listener: (event: FileUploadedEvent<Content>)=>void) {
             this.imageUploader.onFileUploaded(listener);
         }
 
-        unImageUploaded(listener: (event: api.ui.uploader.FileUploadedEvent) => void) {
+        unImageUploaded(listener: (event: FileUploadedEvent<Content>)=>void) {
             this.imageUploader.unFileUploaded(listener);
+        }
+
+        onUploadCompleted(listener: (event: FileUploadCompleteEvent<Content>) => void) {
+            this.imageUploader.onUploadCompleted(listener)
+        }
+
+        unUploadCompleted(listener: (event: FileUploadCompleteEvent<Content>) => void) {
+            this.imageUploader.unUploadCompleted(listener)
+        }
+
+        onUploadReset(listener: () => void) {
+            this.imageUploader.onUploadReset(listener);
+        }
+
+        unUploadReset(listener: () => void) {
+            this.imageUploader.unUploadReset(listener);
         }
 
     }
