+++ modules/wem-admin-ui/src/main/resources/web/admin/common/js/content/form/inputtype/image/_module.ts
@@ -6,6 +6,5 @@
 ///<reference path='ImageUploadDialog.ts' />
 ///<reference path='SelectionToolbar.ts' />
 ///<reference path='FocusChangedEvent.ts' />
-///<reference path='ImageSelectorUploadDialog.ts' />
 ///<reference path='ImageSelectorDisplayValue.ts' />
 ///<reference path='ImageSelectorViewer.ts' />
