+++ modules/wem-admin-ui/src/main/resources/web/admin/common/js/ui/uploader/_module.ts
@@ -3,5 +3,5 @@
 ///<reference path='FileUploadProgressEvent.ts' />
 ///<reference path='FileUploadedEvent.ts' />
 ///<reference path='FileUploadCompleteEvent.ts' />
+///<reference path='Uploader.ts' />
 ///<reference path='FileUploader.ts' />
-///<reference path='ImageUploader.ts' />
