+++ modules/wem-admin-ui/src/main/resources/web/admin/common/styles/api/ui/file-uploader.less
@@ -14,10 +14,10 @@
   .progress-bar {
     width: 100%;
     margin: 25px 0;
+    height: 32px;
     //.linear-gradient(top, #444, #666);
 
     .progress-indicator {
-      height: 32px;
       background-color: #8db800;
       //.linear-gradient(top, #a7d200, #8db800);
     }
